Update 02_basicTextTags.html
So, there are some phrases that don't have periods at the end of them, and I wasn't sure if this was a part of the formatting or not, so I left a lot of them after pointing out my first two suggestions. I think they're mostly the ones that end with <symbol here>.
Same with capitalising some phrases; not sure if you want to capitalise them, but lines 136 to 159 exemplifies what I mean. :)